requirements.txt
streamlit
requests
Pillow
google-genai
huggingface_hub